pyproject.toml
[project]
name = "fadegoblin"
version = "0.1.0"
description = "A Bluesky bot that posts unhinged sports betting parlays."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.65",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
bot = "src.main:main"

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]